Add test case for split plugin
Create feature file for split plugin
Temporary fix for metadata bug in split plugin
Put all plugin step definitions in same  for code reuse
